Rename caching package
