+++ stdlib/ballerina-http/src/main/ballerina/http/http_cache.bal
@@ -16,12 +16,12 @@
 
 package ballerina.http;
 
-import ballerina/caching;
+import ballerina/cache;
 
 type HttpCache object {
 
     private {
-        caching:Cache cache;
+        cache:Cache cache;
         CachingPolicy policy = CACHE_CONTROL_AND_VALIDATORS;
         boolean isShared;
     }
@@ -133,7 +133,7 @@ type HttpCache object {
 
 function createHttpCache (string name, CacheConfig cacheConfig) returns HttpCache {
     HttpCache httpCache = new;
-    caching:Cache backingCache = new(expiryTimeMillis = cacheConfig.expiryTimeMillis, capacity = cacheConfig.capacity,
+    cache:Cache backingCache = new(expiryTimeMillis = cacheConfig.expiryTimeMillis, capacity = cacheConfig.capacity,
                                      evictionFactor = cacheConfig.evictionFactor);
     httpCache.cache = backingCache;
     httpCache.policy = cacheConfig.policy;
@@ -151,7 +151,7 @@ function isCacheableStatusCode (int statusCode) returns boolean {
            statusCode == NOT_IMPLEMENTED_501;
 }
 
-function addEntry (caching:Cache cache, string key, Response inboundResponse) {
+function addEntry (cache:Cache cache, string key, Response inboundResponse) {
     try {
         var existingResponses = cache.get(key);
         match <Response[]>existingResponses {
